multiple scans, choice in picture stack order, auto-stack with helicon focus, height correction, multiple exposures
The user can now scan several areas in sequence

The user can now choose to take stacks from the bottom up or from the top down.
/!\ Stacks taken from the top down will take longer to make :
A vertical movement will first take place in order to reduce horizontal drifting.

After scanning, pictures focus-stack automatically and the resulting pictures are stored in a separate folder. The stacked pictures are now named after the folder they are made from.

The base height of a stack is now adjusted by a correction term calculated from the position of the FL, BR and BL points. If the BL is not defined, the correction will still be calculated, but may be less accurate.

photo_test.py allows the user to scan a subject with a set of different exposure durations. Upon taking a stack, several pictures will be taken for every height, each corresponding to a different exposure time.
Note: because of the camera’s exposure updating rate, 300ms must pass between each picture.